libft + Makefile + gitignore

+++ ft_calloc.c
@@ -6,15 +6,15 @@
 /*   By: rmarcano <rmarcano@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/16 11:48:23 by rmarcano          #+#    #+#             */
-/*   Updated: 2024/01/16 13:45:46 by rmarcano         ###   ########.fr       */
+/*   Updated: 2024/01/17 12:17:30 by rmarcano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 void	*ft_calloc(size_t count, size_t size)
 {
-	int	*arr;
-	int	i;
+	int		*arr;
+	size_t	i;
 
 	arr = malloc(count * size);
 	i = 0;
